10461-story: get rid of TrialsSessions folder, and add an easier-to-understand sort function
