+++ web-client/src/views/AppComponentPublic.tsx
@@ -12,7 +12,7 @@ import { PublicCaseDetail } from './Public/PublicCaseDetail';
 import { PublicPrintableDocketRecord } from './Public/PublicPrintableDocketRecord';
 import { PublicSearch } from './Public/PublicSearch';
 import { PublicTrialSessionDetails } from '@web-client/views/Public/TrialSessions/PublicTrialSessionDetails';
-import { PublicTrialSessions } from '@web-client/views/Public/TrialsSessions/PublicTrialSessions';
+import { PublicTrialSessions } from '@web-client/views/Public/TrialSessions/PublicTrialSessions';
 import { TodaysOpinions } from './Public/TodaysOpinions';
 import { TodaysOrders } from './Public/TodaysOrders';
 import { UsaBanner } from './UsaBanner';
