+++ web-client/src/views/Public/TrialSessions/PublicTrialSessionDetails.tsx
@@ -13,6 +13,7 @@ import { TrialSessionDetailsHeader } from '@web-client/views/TrialSessionDetails
 import { connect } from '@web-client/presenter/shared.cerebral';
 import { state } from '@web-client/presenter/app-public.cerebral';
 import React from 'react';
+import classNames from 'classnames';
 
 export const PublicTrialSessionDetails = connect(
   {
@@ -158,7 +159,15 @@ function NonMobileOpenCases({
                   <CaseIcons formattedCase={publicCase} />
                 </td>
                 <td>
-                  <CaseLink formattedCase={publicCase} />
+                  <span
+                    className={classNames({
+                      'margin-left-2':
+                        publicCase.inConsolidatedGroup &&
+                        !publicCase.isLeadCase,
+                    })}
+                  >
+                    <CaseLink formattedCase={publicCase} />
+                  </span>
                 </td>
                 <td>{publicCase.caseTitle}</td>
                 <td>
