+++ web-client/src/views/Public/TrialSessions/PublicTrialSessionsRemoteProceedingsCard.tsx
